Use integer imag2mix%gtype to safely use enumerator. Use a subroutine to convert gtype to rcmGtype before interpolating to rcmGtype_mix.

+++ src/base/types/volttypes.F90
@@ -50,7 +50,7 @@ module volttypes
         real(rp), dimension(:,:), allocatable :: latc,lonc
         real(rp), dimension(:,:), allocatable :: fac
         real(rp), dimension(:,:), allocatable :: eden,epre ! add electron density and pressure channels to REMIX.
-        real(rp), dimension(:,:), allocatable :: gtype ! RCM grid info: active, buffer, or outside
+        integer , dimension(:,:), allocatable :: gtype ! RCM grid info: active, buffer, or outside
         
         logical, dimension(:,:), allocatable :: inIMag
     end type imag2Mix_T
